requirements.txt
aiohttp
websockets
google-auth
google-auth-httplib2
certifi
requests
loguru